% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAllStations.R
\name{downloadAllStations}
\alias{downloadAllStations}
\title{Download pluviometric data from all pluviometric stations}
\usage{
downloadAllStations(
  pluvDownFolder,
  start_year,
  end_year,
  station.types = c("AUTO", "MAN"),
  download = T
)
}
\arguments{
\item{pluvDownFolder}{path for the pluviometric data to be stored}

\item{start_year}{starting year for the data download}

\item{end_year}{ending year for the data download}

\item{station.types}{either \code{"MAN"} for data from manual stations,
\code{"AUTO"} for data from automatic stations or \code{c("MAN","AUTO")} for
data from both types of stations}

\item{download}{logical argument. If \code{TRUE}, downloads data for all the
available stations and dates, overwriting the ones previously saved in the path
\code{pluvDownFolder}}
}
\description{
Download pluviometric data from all pluviometric stations
}
\examples{
# TODO some example?
}
