% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{getStationList}
\alias{getStationList}
\title{Search for available pluviometric stations}
\usage{
getStationList()
}
\value{
a \code{list} containing information about available locations
(id, type and name)
}
\description{
Search for available pluviometric stations
}
