% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStationLocationAltitude.R
\name{getStationLocationAltitude}
\alias{getStationLocationAltitude}
\title{Obtain location and altitude for a given station}
\usage{
getStationLocationAltitude(st)
}
\arguments{
\item{st}{a vector containing (ID, type, station.name)}
}
\value{
a vector containing (Station.ID,Station.Name,Altitude,UTM.X.Coordinate,UTM.Y.Coordinate)
}
\description{
Obtain location and altitude for a given station
}
