% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_tps.R
\name{interpolatePrecTps}
\alias{interpolatePrecTps}
\title{Interpolate precipitation data for satelite-like rasters using Tps}
\usage{
interpolatePrecTps(pluvDownFolder, dates, type = c("AUTO", "MAN"), template)
}
\arguments{
\item{pluvDownFolder}{path to the folder containing the data}

\item{dates}{vector of dates to interpolate}

\item{type}{type of stations to use for the Tps model, either \code{"AUTO"},\code{"MAN"}, or
\code{c("AUTO","MAN")}}

\item{template}{\code{RasterLayer} to use as a template for prediction}
}
\description{
Interpolate precipitation data for satelite-like rasters using Tps
}
