% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nav.utm.lsat}
\alias{nav.utm.lsat}
\title{Reference raster with Landsat-like resolution}
\format{
The \code{RasterStack} coordinates are in the UTM projection.

\describe{
  \item{resolution}{has an approximate pixel resolution of 30m x 30m}.
  \item{size}{5308 rows by 6709 columns}.
}
}
\description{
Reference \code{RasterLayer} with Landsat-like resolution for a bounding box
over the region of Navarre
}
\keyword{data}
