% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nav.utm.mod}
\alias{nav.utm.mod}
\title{Reference raster with MODIS-like resolution}
\format{
The \code{RasterStack} coordinates are in the UTM projection.

\describe{
  \item{resolution}{has an approximate pixel resolution of 500m x 500m}.
  \item{size}{352 rows by 336 columns}.
}
}
\description{
Reference \code{RasterLayer} with MODIS-like resolution for a bounding box
over the region of Navarre
}
\keyword{data}
