% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nav.utm.sen.10m}
\alias{nav.utm.sen.10m}
\title{Reference raster with Sentinel-like resolution}
\format{
The \code{RasterStack} coordinates are in the UTM projection.

\describe{
  \item{resolution}{has an approximate pixel resolution of 10m x 10m}.
  \item{size}{15593 rows by 14459 columns}.
}
}
\description{
Reference \code{RasterLayer} with Sentinel-like resolution for a bounding box
over the region of Navarre
}
\keyword{data}
